+++ crates/crb-agent/src/address_ext.rs
@@ -2,42 +2,42 @@ use crate::address::Address;
 use crate::agent::Agent;
 use crate::context::Context;
 
-pub trait AddressFor<A: Agent> {
-    fn address(&self) -> Address<A>;
+pub trait ToAddress<A: Agent> {
+    fn to_address(&self) -> Address<A>;
 }
 
-impl<A: Agent> AddressFor<A> for Address<A> {
-    fn address(&self) -> Address<A> {
+impl<A: Agent> ToAddress<A> for Address<A> {
+    fn to_address(&self) -> Address<A> {
         self.clone()
     }
 }
 
-impl<A: Agent> AddressFor<A> for &mut Address<A> {
-    fn address(&self) -> Address<A> {
+impl<A: Agent> ToAddress<A> for &mut Address<A> {
+    fn to_address(&self) -> Address<A> {
         (*self).clone()
     }
 }
 
-impl<A: Agent> AddressFor<A> for &Address<A> {
-    fn address(&self) -> Address<A> {
+impl<A: Agent> ToAddress<A> for &Address<A> {
+    fn to_address(&self) -> Address<A> {
         (*self).clone()
     }
 }
 
-impl<A: Agent> AddressFor<A> for Context<A> {
-    fn address(&self) -> Address<A> {
-        Context::address(self).clone()
+impl<A: Agent> ToAddress<A> for Context<A> {
+    fn to_address(&self) -> Address<A> {
+        self.address().clone()
     }
 }
 
-impl<A: Agent> AddressFor<A> for &Context<A> {
-    fn address(&self) -> Address<A> {
-        Context::address(self).clone()
+impl<A: Agent> ToAddress<A> for &Context<A> {
+    fn to_address(&self) -> Address<A> {
+        self.address().clone()
     }
 }
 
-impl<A: Agent> AddressFor<A> for &mut Context<A> {
-    fn address(&self) -> Address<A> {
-        Context::address(self).clone()
+impl<A: Agent> ToAddress<A> for &mut Context<A> {
+    fn to_address(&self) -> Address<A> {
+        self.address().clone()
     }
 }